projects/capstone/starter/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.5.3
Flask-OAuthlib==0.9.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
jose==1.0.0
migrate==0.3.7
oauthlib==2.1.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
psycopg2-pool==1.1
py==1.10.0
python-dotenv==0.19.0
python-jose==3.3.0
requests==2.26.0
requests-oauthlib==1.1.0
SQLAlchemy==1.4.20
virtualenv==20.4.7
virtualenvwrapper-win==1.2.6
Werkzeug==2.0.1